Moved it to a different file to keep it more organized
