+++ Frontend/src/components/AdminSettings/AdminSettings.css
@@ -33,16 +33,6 @@
   margin: 0 2em 1em 0;
 }
 
-.color-availability-mark {
-  border: none;
-  position: absolute;
-  color: gray;
-  align-self: baseline;
-  padding-top: 0.1em;
-  font-size: 2em;
-  padding-left: 0.5em;
-}
-
 #FFFFFF {
   color: black;
 }
